chainlit-sample-db-chatprofile/requirements.txt
chainlit>=2.5.0
openai>=1.84.0
python-dotenv>=1.0.0
asyncpg>=0.30.0
SQLAlchemy>=2.0.40
psycopg2-binary>=2.9.10
greenlet>=3.2.0
